refactor: モックファイルをソース構造に対応するディレクトリに移動
- tests/integration/.../mocks/ → tests/frameworks-and-drivers/browser/ChromeTabsGateway/
- テストファイルのインポートパスを更新
- docs/coding-standards/tests/common-rule.md にモック配置ルールを追加
